liver_tumor_ai/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
python-dotenv
markdown
